fix: 与钱有关的东西全部转化为以 '分' 为单位的 Int

+++ src/main/java/fun/steven/bookstore/dao/OrderDao.java
@@ -52,7 +52,7 @@ public class OrderDao implements IOrderDao {
         order.setOrderItems(orderItems);
 
         /* 计算总价格 */
-        Double totalPrice = selectedCartItems.stream().mapToDouble(
+        Integer totalPrice = selectedCartItems.stream().mapToInt(
                 item -> item.getBook().getPrice() * item.getQuantity()).sum();
         order.setTotalPrice(totalPrice);
